[BugFix] Forgot to run drop table task
