+++ fe/fe-core/src/main/java/com/starrocks/lake/LakeTable.java
@@ -11,13 +11,10 @@ import com.starrocks.catalog.KeysType;
 import com.starrocks.catalog.MaterializedIndex;
 import com.starrocks.catalog.MaterializedView;
 import com.starrocks.catalog.OlapTable;
-import com.starrocks.catalog.Partition;
 import com.starrocks.catalog.PartitionInfo;
 import com.starrocks.catalog.Table;
 import com.starrocks.catalog.TableIndexes;
 import com.starrocks.catalog.TableProperty;
-import com.starrocks.catalog.Tablet;
-import com.starrocks.catalog.TabletInvertedIndex;
 import com.starrocks.common.DdlException;
 import com.starrocks.common.io.DeepCopy;
 import com.starrocks.common.io.Text;
@@ -33,9 +30,7 @@ import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.Collection;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 /**
  * Metadata for StarRocks lake table
@@ -136,47 +131,6 @@ public class LakeTable extends OlapTable {
     @Override
     public Runnable delete(boolean replay) {
         GlobalStateMgr.getCurrentState().getLocalMetastore().onEraseTable(this);
-        return replay ? null : new DeleteTableTask(this, replay);
-    }
-
-    private static class DeleteTableTask implements Runnable {
-        private final LakeTable table;
-        private final boolean replay;
-
-        DeleteTableTask(LakeTable table, boolean replay) {
-            this.table = table;
-            this.replay = replay;
-        }
-
-        @Override
-        public void run() {
-            // inverted index
-            TabletInvertedIndex invertedIndex = GlobalStateMgr.getCurrentInvertedIndex();
-            Collection<Partition> allPartitions = table.getAllPartitions();
-            for (Partition partition : allPartitions) {
-                for (MaterializedIndex index : partition.getMaterializedIndices(MaterializedIndex.IndexExtState.ALL)) {
-                    for (Tablet tablet : index.getTablets()) {
-                        invertedIndex.deleteTablet(tablet.getId());
-                    }
-                }
-            }
-
-            if (replay) {
-                return;
-            }
-
-            // drop all replicas
-            Set<Long> tabletIds = new HashSet<>();
-            for (Partition partition : table.getAllPartitions()) {
-                List<MaterializedIndex> allIndices = partition.getMaterializedIndices(MaterializedIndex.IndexExtState.ALL);
-                for (MaterializedIndex materializedIndex : allIndices) {
-                    for (Tablet tablet : materializedIndex.getTablets()) {
-                        tabletIds.add(tablet.getId());
-                    }
-                }
-            }
-            GlobalStateMgr.getCurrentState().getShardManager().getShardDeleter().addUnusedShardId(tabletIds);
-            GlobalStateMgr.getCurrentState().getEditLog().logAddUnusedShard(tabletIds);
-        }
+        return replay ? null : new DeleteTableTask(this);
     }
 }
